Print Player Coordinates
Set up a basic function that will print the Hero's coordinates to the terminal. Made this occur through the Update function set to a 1 second delay on each print. May use the Update method to run a movement timer on the Minotaur so they aren't moving every single frame. I will probably make it so the Minotaur has a movement opportunity, where it has a chance in 20 to move, similar to that of the AI in Five Nights at Freddy's.