Basic A* Function and Printing Taxicab Distance
set up basic A* pathfinding systems. Created the Open and Closed lists for the tiles, set the boss tile as the start tile and made that have an F Cost of zero.
Additionally added printing the current taxicab distance between the Hero and the Boss on the same timing their coordinates are printed. This may be commented out before final implementation if I choose to use the console for any gameplay elements.