A* Pathfinding Implementation
Resets per-tile state, uses the boss tile as the start and Hero as the goal, uses Open and Closed Set loop, updates the g/f costs, stores the parent.

Next to do is reconstructing the path and moving the Minotaur one tile along the path every time the update is called.